+++ src/shared/options.ts
@@ -83,6 +83,7 @@ export interface LanguageServerOptions {
     readonly useServerGC: boolean;
     readonly reportInformationAsHint: boolean;
     readonly environmentVariables: { [key: string]: string };
+    readonly sourceGeneratorExecution: string;
 }
 
 export interface RazorOptions {
@@ -421,6 +422,9 @@ class LanguageServerOptionsImpl implements LanguageServerOptions {
     public get environmentVariables() {
         return readOption<{ [key: string]: string }>('dotnet.server.environmentVariables', {});
     }
+    public get sourceGeneratorExecution() {
+        return readOption<string>('dotnet.server.sourceGeneratorExecution', 'Balanced');
+    }
 }
 
 class RazorOptionsImpl implements RazorOptions {
@@ -517,5 +521,6 @@ export const LanguageServerOptionsThatTriggerReload: ReadonlyArray<keyof Languag
     'useServerGC',
     'reportInformationAsHint',
     'environmentVariables',
+    'sourceGeneratorExecution',
     'crashDumpPath',
 ];