+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 * Enable and switch to `Balanced` source generator execution (PR: [#8970](https://github.com/dotnet/vscode-csharp/pull/8970))
   * `Balanced` mode improves language server performance by only running source generators on explicit actions like file save, build task execution, or the `csharp.rerunSourceGenerators` command.  This is in contrast to the previous default, `Automatic`, which ran source generators on every keystroke.
   * The `dotnet.server.sourceGeneratorExecution` option allows you to switch between `Balanced` and `Automatic` source generator execution (requires restart).
+* Update Roslyn to 5.5.0-2.26103.6 (PR: [#8956](https://github.com/dotnet/vscode-csharp/pull/8956))
+  * Fix workspace search always returning no results for first query (PR: [#82276](https://github.com/dotnet/roslyn/pull/82276))
+  * Improve elimination of redundant evaluations during pattern matching operation (PR: [#82142](https://github.com/dotnet/roslyn/pull/82142))
+  * Limit the FileSystemWatchers to one per drive root for DefaultFileChangeWatcher (PR: [#82211](https://github.com/dotnet/roslyn/pull/82211))
+  * Reduce allocations during analysis result creation. (PR: [#82139](https://github.com/dotnet/roslyn/pull/82139))
+  * Set isReferenceAssembly to false if we found the implementation assembly (PR: [#82242](https://github.com/dotnet/roslyn/pull/82242))
+  * Remove console title change from language server Program.cs (PR: [#82229](https://github.com/dotnet/roslyn/pull/82229))
 
 # 2.121.x
 * Use the package.json to gather all defined settings (PR: [#8954](https://github.com/dotnet/vscode-csharp/pull/8954))